Generalize over EVMLA data
