+++ src/eravm/context/address_space.rs
@@ -2,7 +2,7 @@
 //! The address space aliases.
 //!
 
-use crate::context::address_space::IAddressSpace;
+use crate::context::traits::address_space::IAddressSpace;
 
 ///
 /// The address space aliases.
@@ -28,38 +28,6 @@ pub enum AddressSpace {
 impl IAddressSpace for AddressSpace {
     fn stack() -> Self {
         Self::Stack
-    }
-
-    fn heap() -> Self {
-        Self::Heap
-    }
-
-    fn aux_heap() -> Self {
-        Self::HeapAuxiliary
-    }
-
-    fn calldata() -> Self {
-        panic!("Only available for EVM");
-    }
-
-    fn return_data() -> Self {
-        panic!("Only available for EVM");
-    }
-
-    fn generic() -> Self {
-        Self::Generic
-    }
-
-    fn code() -> Self {
-        Self::Code
-    }
-
-    fn storage() -> Self {
-        Self::Storage
-    }
-
-    fn transient_storage() -> Self {
-        Self::TransientStorage
     }
 }
