+++ src/eravm/context/mod.rs
@@ -22,7 +22,6 @@ use std::rc::Rc;
 use inkwell::types::BasicType;
 use inkwell::values::BasicValue;
 
-use crate::context::address_space::IAddressSpace;
 use crate::context::attribute::Attribute;
 use crate::context::code_type::CodeType;
 use crate::context::function::declaration::Declaration as FunctionDeclaration;
@@ -916,99 +915,6 @@ where
         Ok(())
     }
 
-    fn build_alloca<T>(&self, r#type: T, name: &str) -> Pointer<'ctx, Self::AddressSpace>
-    where
-        T: BasicType<'ctx> + Clone + Copy,
-    {
-        let pointer = self.builder.build_alloca(r#type, name);
-        self.basic_block()
-            .get_last_instruction()
-            .expect("Always exists")
-            .set_alignment(era_compiler_common::BYTE_LENGTH_FIELD as u32)
-            .expect("Alignment is valid");
-        Pointer::new(r#type, Self::AddressSpace::stack(), pointer)
-    }
-
-    fn build_load(
-        &self,
-        pointer: Pointer<'ctx, Self::AddressSpace>,
-        name: &str,
-    ) -> inkwell::values::BasicValueEnum<'ctx> {
-        let value = self.builder.build_load(pointer.r#type, pointer.value, name);
-
-        let alignment = if Self::AddressSpace::stack() == pointer.address_space {
-            era_compiler_common::BYTE_LENGTH_FIELD
-        } else {
-            era_compiler_common::BYTE_LENGTH_BYTE
-        };
-
-        self.basic_block()
-            .get_last_instruction()
-            .expect("Always exists")
-            .set_alignment(alignment as u32)
-            .expect("Alignment is valid");
-        value
-    }
-
-    fn build_store<V>(&self, pointer: Pointer<'ctx, Self::AddressSpace>, value: V)
-    where
-        V: BasicValue<'ctx>,
-    {
-        let instruction = self.builder.build_store(pointer.value, value);
-
-        let alignment = if Self::AddressSpace::stack() == pointer.address_space {
-            era_compiler_common::BYTE_LENGTH_FIELD
-        } else {
-            era_compiler_common::BYTE_LENGTH_BYTE
-        };
-
-        instruction
-            .set_alignment(alignment as u32)
-            .expect("Alignment is valid");
-    }
-
-    fn build_gep<T>(
-        &self,
-        pointer: Pointer<'ctx, Self::AddressSpace>,
-        indexes: &[inkwell::values::IntValue<'ctx>],
-        element_type: T,
-        name: &str,
-    ) -> Pointer<'ctx, Self::AddressSpace>
-    where
-        T: BasicType<'ctx>,
-    {
-        let value = unsafe {
-            self.builder
-                .build_gep(pointer.r#type, pointer.value, indexes, name)
-        };
-        Pointer::new(element_type, pointer.address_space, value)
-    }
-
-    fn build_conditional_branch(
-        &self,
-        comparison: inkwell::values::IntValue<'ctx>,
-        then_block: inkwell::basic_block::BasicBlock<'ctx>,
-        else_block: inkwell::basic_block::BasicBlock<'ctx>,
-    ) {
-        if self.basic_block().get_terminator().is_some() {
-            return;
-        }
-
-        self.builder
-            .build_conditional_branch(comparison, then_block, else_block);
-    }
-
-    fn build_unconditional_branch(
-        &self,
-        destination_block: inkwell::basic_block::BasicBlock<'ctx>,
-    ) {
-        if self.basic_block().get_terminator().is_some() {
-            return;
-        }
-
-        self.builder.build_unconditional_branch(destination_block);
-    }
-
     fn build_call(
         &self,
         function: FunctionDeclaration<'ctx>,
@@ -1106,115 +1012,6 @@ where
             self.build_store(return_pointer, return_value);
         }
         return_pointer.map(|pointer| self.build_load(pointer, "invoke_result"))
-    }
-
-    fn build_memcpy(
-        &self,
-        function: FunctionDeclaration<'ctx>,
-        destination: Pointer<'ctx, AddressSpace>,
-        source: Pointer<'ctx, AddressSpace>,
-        size: inkwell::values::IntValue<'ctx>,
-        name: &str,
-    ) {
-        let call_site_value = self.builder.build_indirect_call(
-            function.r#type,
-            function.value.as_global_value().as_pointer_value(),
-            &[
-                destination.value.as_basic_value_enum().into(),
-                source.value.as_basic_value_enum().into(),
-                size.as_basic_value_enum().into(),
-                self.bool_type().const_zero().as_basic_value_enum().into(),
-            ],
-            name,
-        );
-
-        call_site_value.set_alignment_attribute(inkwell::attributes::AttributeLoc::Param(0), 1);
-        call_site_value.set_alignment_attribute(inkwell::attributes::AttributeLoc::Param(1), 1);
-    }
-
-    fn build_return(&self, value: Option<&dyn BasicValue<'ctx>>) {
-        if self.basic_block().get_terminator().is_some() {
-            return;
-        }
-
-        self.builder.build_return(value);
-    }
-
-    fn build_unreachable(&self) {
-        if self.basic_block().get_terminator().is_some() {
-            return;
-        }
-
-        self.builder.build_unreachable();
-    }
-
-    fn bool_const(&self, value: bool) -> inkwell::values::IntValue<'ctx> {
-        self.bool_type().const_int(u64::from(value), false)
-    }
-
-    fn integer_const(&self, bit_length: usize, value: u64) -> inkwell::values::IntValue<'ctx> {
-        self.integer_type(bit_length).const_int(value, false)
-    }
-
-    fn field_const(&self, value: u64) -> inkwell::values::IntValue<'ctx> {
-        self.field_type().const_int(value, false)
-    }
-
-    fn field_undef(&self) -> inkwell::values::IntValue<'ctx> {
-        self.field_type().get_undef()
-    }
-
-    fn field_const_str_dec(&self, value: &str) -> inkwell::values::IntValue<'ctx> {
-        self.field_type()
-            .const_int_from_string(value, inkwell::types::StringRadix::Decimal)
-            .unwrap_or_else(|| panic!("Invalid string constant `{value}`"))
-    }
-
-    fn field_const_str_hex(&self, value: &str) -> inkwell::values::IntValue<'ctx> {
-        self.field_type()
-            .const_int_from_string(
-                value.strip_prefix("0x").unwrap_or(value),
-                inkwell::types::StringRadix::Hexadecimal,
-            )
-            .unwrap_or_else(|| panic!("Invalid string constant `{value}`"))
-    }
-
-    fn void_type(&self) -> inkwell::types::VoidType<'ctx> {
-        self.llvm.void_type()
-    }
-
-    fn bool_type(&self) -> inkwell::types::IntType<'ctx> {
-        self.llvm.bool_type()
-    }
-
-    fn byte_type(&self) -> inkwell::types::IntType<'ctx> {
-        self.llvm
-            .custom_width_int_type(era_compiler_common::BIT_LENGTH_BYTE as u32)
-    }
-
-    fn integer_type(&self, bit_length: usize) -> inkwell::types::IntType<'ctx> {
-        self.llvm.custom_width_int_type(bit_length as u32)
-    }
-
-    fn field_type(&self) -> inkwell::types::IntType<'ctx> {
-        self.llvm
-            .custom_width_int_type(era_compiler_common::BIT_LENGTH_FIELD as u32)
-    }
-
-    fn array_type<T>(&self, element_type: T, length: usize) -> inkwell::types::ArrayType<'ctx>
-    where
-        T: BasicType<'ctx>,
-    {
-        element_type.array_type(length as u32)
-    }
-
-    fn structure_type<T>(&self, field_types: &[T]) -> inkwell::types::StructType<'ctx>
-    where
-        T: BasicType<'ctx>,
-    {
-        let field_types: Vec<inkwell::types::BasicTypeEnum<'ctx>> =
-            field_types.iter().map(T::as_basic_type_enum).collect();
-        self.llvm.struct_type(field_types.as_slice(), false)
     }
 
     fn set_solidity_data(&mut self, data: Self::SolidityData) {
