+++ src/eravm/evm/calldata.rs
@@ -4,7 +4,6 @@
 
 use inkwell::types::BasicType;
 
-use crate::context::address_space::IAddressSpace;
 use crate::context::pointer::Pointer;
 use crate::context::IContext;
 use crate::eravm::context::address_space::AddressSpace;
@@ -66,7 +65,7 @@ where
 {
     let destination = Pointer::new_with_offset(
         context,
-        AddressSpace::heap(),
+        AddressSpace::Heap,
         context.byte_type(),
         destination_offset,
         "calldata_copy_destination_pointer",
